Question 1
This is the actual source code, the data for the pictures were located in the assets folder, which becuase can get very big, i am not able to load it to the gihub. In a production world, i would replace those with a blob stroage, and inject the data.json with remote access to the assets.